pyproject.toml
[tool.poetry]
name = "simpledrill"
version = "0.1.0"
description = ""
authors = ["Larisa Shigareva <thelarisa81@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.6"
python-dotenv = "^0.15.0"
django-threadlocals = "^0.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
